#384 Duplicated classes
