+++ core/roboconf-plugin-script/pom.xml
@@ -60,6 +60,12 @@
   			<version>${project.version}</version>
   			<scope>provided</scope>
 		</dependency>
+		
+		<dependency>
+			<groupId>com.github.spullara.mustache.java</groupId>
+			<artifactId>compiler</artifactId>
+			<version>0.8.14</version>
+		</dependency>
 
 		<dependency>
 			<groupId>net.roboconf</groupId>
