+++ core/roboconf-plugin-script/src/test/resources/importTemplate.mustache
