+++ core/roboconf-plugin-script/src/test/resources/instanceTemplate.mustache
